CI on MySQL: nationality as a country list, and driver-independent tests
- Passenger nationality was a free-text box writing to a char(2) column.
  SQLite stored «Ελληνική»; MySQL refused it with a 500 at checkout. It is
  now a list of ISO 3166 countries named in the guest's language (ICU, via
  intl), validated at checkout and normalised on the /g/ details page.
- CatalogueWebhookTest: on the Redis queue the trip's own creation job holds
  the unique lock, silencing the save under test; release it after setup,
  and run the end-to-end case on the sync queue.
- sum() is a string on MySQL: cast in two tests.
- A MySQL JSON column reorders object keys: compare itinerary stops with
  toEqual.
- Pint: a stray blank line in bootstrap/app.php.

+++ tests/Feature/Booking/RemoveGuestsFromBookingTest.php
@@ -91,7 +91,8 @@ it('takes one person off at their own price, gives the seat back and refunds the
             ->and($fresh->pax_capacity_total)->toBe(3)
             ->and($fresh->pax_breakdown[0]['qty'])->toBe(2)
             ->and(Departure::query()->findOrFail($booking->departure_id)->seats_sold)->toBe(3)
-            ->and(Payment::query()->where('booking_id', $booking->getKey())->where('kind', PaymentKind::Refund->value)->sum('amount_cents'))->toBe(5000)
+            // `sum()` is a string on MySQL and a number on SQLite.
+            ->and((int) Payment::query()->where('booking_id', $booking->getKey())->where('kind', PaymentKind::Refund->value)->sum('amount_cents'))->toBe(5000)
             ->and($fresh->paid_cents)->toBe(13000)
             ->and($fresh->status)->toBe(BookingStatus::Confirmed);
     });
